requirements.txt
torch==2.0.1
hydra-core==1.3.2
scipy==1.11.2
tqdm==4.66.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
